로그인 및 예외처리 추가

+++ src/main/java/com/tistory/hitomis/springbootthymeleafboard/controller/BoardController.java
@@ -1,17 +1,21 @@
 package com.tistory.hitomis.springbootthymeleafboard.controller;
 
 import com.tistory.hitomis.springbootthymeleafboard.domain.Board;
+import com.tistory.hitomis.springbootthymeleafboard.domain.Member;
 import com.tistory.hitomis.springbootthymeleafboard.service.BoardService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 
@@ -53,6 +57,11 @@ public class BoardController {
         return "testBoardList"; // jsp 파일 이름
     }
 
+    @ModelAttribute("member")
+    public Member setMember() {
+        return new Member();
+    }
+
     /**
      * 게시판 목록
      *
@@ -61,7 +70,10 @@ public class BoardController {
      * @return
      */
     @RequestMapping("/getBoardList")
-    public String getBoardList(Model model, Board board) {
+    public String getBoardList(@ModelAttribute("member") Member member, Model model, Board board) {
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         List<Board> boardList = boardService.getBoardList(board);
         model.addAttribute("boardList", boardList);
         return "getBoardList";
@@ -74,7 +86,10 @@ public class BoardController {
      * @return
      */
     @RequestMapping("/insertBoardView")
-    public String insertBoardView() {
+    public String insertBoardView(@ModelAttribute("member") Member member) {
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         return "insertBoard";
     }
 
@@ -85,7 +100,10 @@ public class BoardController {
      * @return
      */
     @RequestMapping("/insertBoard")
-    public String insertBoard(Board board) {
+    public String insertBoard(@ModelAttribute("member") Member member, Board board) {
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         boardService.insertBoard(board);
         return "redirect:getBoardList";
     }
@@ -98,7 +116,10 @@ public class BoardController {
      * @return
      */
     @RequestMapping("/getBoard")
-    public String getBoard(Board board, Model model) {
+    public String getBoard(@ModelAttribute("member") Member member, Board board, Model model) {
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         model.addAttribute("board", boardService.getBoard((board)));
         return "getBoard";
     }
@@ -110,20 +131,26 @@ public class BoardController {
      * @return
      */
     @RequestMapping("/updateBoard")
-    public String updateBoard(Board board) {
+    public String updateBoard(@ModelAttribute("member") Member member, Board board) {
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         boardService.updateBoard(board);
         return "forward:getBoardList";
     }
 
     /**
      * 글 삭제 처리 후 목록으로 이동
+     *
      * @param board
      * @return
      */
     @RequestMapping("/deleteBoard")
-    public String deleteBoard(Board board) {
+    public String deleteBoard(@ModelAttribute("member") Member member, Board board) {
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         boardService.deleteBoard(board);
         return "forward:getBoardList";
     }
-
 }
